Thu, Jan 11, 2024, 9:54 AM +00:00

+++ tideman/tideman.c
@@ -224,7 +224,7 @@ void print_winner(void)
     {
         for (int i = 0; i < candidate_count; i++)
         {
-            losses[j] += (int) locked[i][j];
+            losses[j] += locked[i][j];
             printf("%d ", locked[i][j]);
         }
         printf("Candidate %s has lost %d times\n", candidates[j], losses[j]);
